+++ runnables/train_msm.py
@@ -10,7 +10,7 @@ from src.models.msm import MSM
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
-torch.set_default_dtype(torch.double)
+torch.set_default_dtype(torch.float32)
 
 
 @hydra.main(config_name=f'config.yaml', config_path='../config/')
@@ -110,4 +110,3 @@ def main(args: DictConfig):
 
 if __name__ == "__main__":
     main()
-
